# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at <a href=\"http://swagger.io\">http://swagger.io</a> or on irc.freenode.net, #swagger.  For this sample, you can use the api key \"special-key\" to test the authorization filters",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "email": "apiteam@wordnik.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:1337",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/film": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Affiche la liste des films.",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "default": {
                        "description": "An error has been occured"
                    }
                }
            },
            "post": {
                "tags": [
                    "film"
                ],
                "summary": "Store a newly created film in storage.",
                "operationId": "postFilm",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "nom",
                        "in": "formData",
                        "description": "Name of the discount",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date_debut",
                        "in": "formData",
                        "description": "Begin of the discount",
                        "required": false,
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "date_fin",
                        "in": "formData",
                        "description": "End of the discount",
                        "required": false,
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "pourcentage_reduction",
                        "in": "formData",
                        "description": "Percentage of the discount",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Reduction"
                            }
                        }
                    },
                    "404": {
                        "description": "Woops! The informations can't be validated!",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Reduction"
                            }
                        }
                    },
                    "default": {
                        "description": "An error has been occured"
                    }
                }
            }
        },
...
    },
...
    "definitions": {
        "Reduction": {
            "required": [
                "id_reduction"
            ],
            "properties": {
                "id_reduction": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "nom": {
                    "type": "string",
                    "format": "string",
                    "default": "r\u00e9duction"
                },
                "date_debut": {
                    "type": "date",
                    "format": "date",
                    "default": "2000-01-01"
                },
                "date_fin": {
                    "type": "date",
                    "format": "date",
                    "default": "2000-01-02"
                },
                "pourcentage_reduction": {
                    "type": "number",
                    "format": "int64",
                    "default": 0
                }
            },
            "xml": {
                "name": "Film"
            }
        }
    },
...
}